+++ src/gort/recipes/calibrations.py
@@ -16,11 +16,9 @@ from typing import Sequence
 import numpy
 from astropy.time import Time
 
-from sdsstools import get_sjd
-
 from gort.enums import ErrorCode, Event
 from gort.exceptions import GortError
-from gort.tools import decap, get_ephemeris_summary, redis_client_sync
+from gort.tools import decap, get_ephemeris_summary
 
 from .base import BaseRecipe
 
@@ -364,16 +362,8 @@ class LongTermCalibrations(BaseRecipe):
 
     name = "long_term_calibrations"
 
-    async def recipe(self, biases: bool | None = None):
-        """Runs the calibration sequence.
-
-        Parameters
-        ----------
-        biases
-            If :obj:`True`, takes a sequence of bias frames. If :obj:`None`, checks
-            if the Overwatcher has taken the biases today and won't retake them if so.
-
-        """
+    async def recipe(self):
+        """Runs the calibration sequence."""
 
         config = self.gort.config["recipes"][self.name]
 
@@ -384,13 +374,6 @@ class LongTermCalibrations(BaseRecipe):
         arc_exp_times: Sequence[float] = config.get("arc_exp_times", [10, 50])
         if not isinstance(arc_exp_times, Sequence):
             arc_exp_times = [arc_exp_times]
-
-        if biases is None:
-            biases = self.check_biases()
-
-        if biases:
-            self.gort.log.info(f"Taking {n_biases} bias frames.")
-            await self.gort.execute_recipe("bias_sequence", count=n_biases)
 
         self.gort.log.info("Moving telescopes to point to the calibration screen.")
         await self.gort.telescopes.goto_named_position("calibration")
@@ -465,21 +448,10 @@ class LongTermCalibrations(BaseRecipe):
         self.gort.log.debug("Turning off all lamps.")
         await self.gort.nps.calib.all_off()
 
+        # Biases
+        await self.gort.execute_recipe("bias_sequence", count=n_biases)
+
+        # Park the telescopes after the calibrations.
         await self.gort.telescopes.park(disable=False)
 
-    def check_biases(self) -> bool:
-        """Checks whether the biases have been taken tonight."""
-
-        sjd = get_sjd("LCO")
-
-        with redis_client_sync() as redis:
-            try:
-                key = f"overwatcher:calibrations:{sjd}"
-                bias_data = redis.json().get(key, ".bias_sequence")
-            except Exception:
-                return True
-
-            if bias_data is None or "state" not in bias_data:
-                return True
-
-            return bias_data["state"] == "done"  # type: ignore
+        self.gort.log.info("Long-term calibrations completed successfully.")